beer_challenge/requirements.txt
requests
psycopg2-binary
sqlalchemy
pandas